requirements.txt
asgiref==3.5.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.0.12
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
daphne==3.0.2
dj-database-url==0.5.0
Django==4.1.7
django-cors-headers==3.11.0
django-filter==23.1
django-heroku==0.3.1
django-jalali-date==1.0.1
django-object-tools==2.0.0
django-rest-passwordreset==1.3.0
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
jdatetime==4.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysqlclient==2.1.1
packaging==21.3
Pillow==9.1.0
psycopg2==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.0.3
pyOpenSSL==23.1.1
pyparsing==3.0.7
python-dotenv==1.0.0
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==6.0.0
zope.interface==6.0
django-filter==23.1

